Broadcast (broadcast v0.1.0)

Broadcast to multiple social media domains simultaneously through a unified interface.

Functions

post_all(map)

Posts a given status message to both Mastodon and Bluesky platforms.

Parameters

The function takes a map containing the necessary parameters for posting:

:status - a string representing the status message to be posted.

:mastodon_access_token - the access token for the Mastodon account for authentication.

:bluesky_handle - the handle (username) of the Bluesky account to post the status.

:bluesky_password - the password for the Bluesky account for authentication.

Examples

iex> post_all(%{ ...> status: "Hello World!", ...> mastodon_access_token: "your_mastodon_token", ...> bluesky_handle: "your_bluesky_handle", ...> bluesky_password: "your_bluesky_password" ...> })

post_bluesky_status(handle, password, status)

Posts a status update to the Bluesky social network.

Parameters

handle: a string representing the user identifier on the Bluesky platform.

password: a string containing the user password for authentication.

status: a string representing the status update to be posted.

Examples

iex> post_bluesky_status("your_user_handle", "your_password", "Hello, Mastodon!")

post_mastodon_status(access_token, status)

Posts a status update to the Mastodon social media platform.

Parameters

access_token - The bearer token used for authentication with the Mastodon API.

status - The string representing the status update to be posted.

Examples

iex> post_mastodon_status("your_access_token", "Hello, Mastodon!")